+++ server/src-lib/Hasura/GraphQL/Execute/LiveQuery/Plan.hs
@@ -28,21 +28,21 @@ import           Hasura.Prelude
 
 import qualified Data.Aeson.Casing                           as J
 import qualified Data.Aeson.Extended                         as J
-import qualified Data.ByteString                             as B
-import qualified Data.HashSet                                as Set
 import qualified Data.Aeson.TH                               as J
+import qualified Data.ByteString                             as B
 import qualified Data.HashMap.Strict                         as Map
 import qualified Data.HashMap.Strict.InsOrd                  as OMap
+import qualified Data.HashSet                                as Set
 import qualified Data.Text                                   as T
 import qualified Data.UUID.V4                                as UUID
 import qualified Database.PG.Query                           as Q
-import qualified Language.GraphQL.Draft.Syntax               as G
 import qualified Database.PG.Query.PTI                       as PTI
+import qualified Language.GraphQL.Draft.Syntax               as G
 import qualified PostgreSQL.Binary.Encoding                  as PE
 
 import           Control.Lens
-import           Data.UUID                                   (UUID)
 import           Data.Semigroup.Generic
+import           Data.UUID                                   (UUID)
 
 import qualified Hasura.Backends.Postgres.Execute.RemoteJoin as RR
 import qualified Hasura.Backends.Postgres.SQL.DML            as S
@@ -271,8 +271,8 @@ validateVariables pgExecCtx variableValues = do
 -- of a subscription field's AST as we resolve them to SQL expressions.
 data QueryParametersInfo
   = QueryParametersInfo
-  { _qpiReusableVariableValues     :: !(HashMap G.Name PGColumnValue)
-  , _qpiSyntheticVariableValues    :: !(Seq PGColumnValue)
+  { _qpiReusableVariableValues     :: !(HashMap G.Name (ColumnValue 'Postgres))
+  , _qpiSyntheticVariableValues    :: !(Seq (ColumnValue 'Postgres))
   , _qpiReferencedSessionVariables :: !(Set.HashSet SessionVariable)
   -- ^ The session variables that are referenced in the query root fld's AST.
   -- This information is used to determine a cohort's required session
@@ -287,7 +287,7 @@ makeLenses ''QueryParametersInfo
 -- about various parameters of the query along the way.
 resolveMultiplexedValue
   :: (MonadState QueryParametersInfo m)
-  => UnpreparedValue -> m S.SQLExp
+  => UnpreparedValue 'Postgres -> m S.SQLExp
 resolveMultiplexedValue = \case
   UVParameter colVal varM -> do
     varJsonPath <- case fmap PS.getName varM of
@@ -298,7 +298,7 @@ resolveMultiplexedValue = \case
         syntheticVarIndex <- use (qpiSyntheticVariableValues . to length)
         modifying qpiSyntheticVariableValues (|> colVal)
         pure ["synthetic", T.pack $ show syntheticVarIndex]
-    pure $ fromResVars (PGTypeScalar $ pstType $ pcvValue colVal) varJsonPath
+    pure $ fromResVars (PGTypeScalar $ pstType $ cvValue colVal) varJsonPath
   UVSessionVar ty sessVar -> do
     modifying qpiReferencedSessionVariables (Set.insert sessVar)
     pure $ fromResVars ty ["session", sessionVariableToText sessVar]
@@ -337,7 +337,7 @@ data ReusableLiveQueryPlan
   { _rlqpParameterizedPlan        :: !ParameterizedLiveQueryPlan
   , _rlqpRequiredSessionVariables :: !(Set.HashSet SessionVariable)
   , _rlqpSyntheticVariableValues  :: !ValidatedSyntheticVariables
-  , _rlqpQueryVariableTypes       :: HashMap G.Name PGColumnType
+  , _rlqpQueryVariableTypes       :: HashMap G.Name (ColumnType 'Postgres)
   } deriving (Show)
 $(J.deriveToJSON (J.aesonDrop 4 J.snakeCase) ''ReusableLiveQueryPlan)
 
@@ -349,7 +349,7 @@ buildLiveQueryPlan
      )
   => PGExecCtx
   -> UserInfo
-  -> InsOrdHashMap G.Name (SubscriptionRootField UnpreparedValue)
+  -> InsOrdHashMap G.Name (SubscriptionRootField (UnpreparedValue 'Postgres))
   -> m (LiveQueryPlan, Maybe ReusableLiveQueryPlan)
 buildLiveQueryPlan pgExecCtx userInfo unpreparedAST = do
   (preparedAST, QueryParametersInfo{..}) <- flip runStateT mempty $
@@ -374,8 +374,8 @@ buildLiveQueryPlan pgExecCtx userInfo unpreparedAST = do
   -- We need to ensure that the values provided for variables are correct according to Postgres.
   -- Without this check an invalid value for a variable for one instance of the subscription will
   -- take down the entire multiplexed query.
-  validatedQueryVars <- validateVariables pgExecCtx $ fmap pcvValue _qpiReusableVariableValues
-  validatedSyntheticVars <- validateVariables pgExecCtx $ map pcvValue $ toList _qpiSyntheticVariableValues
+  validatedQueryVars <- validateVariables pgExecCtx $ fmap cvValue _qpiReusableVariableValues
+  validatedSyntheticVars <- validateVariables pgExecCtx $ map cvValue $ toList _qpiSyntheticVariableValues
 
   let -- TODO validatedQueryVars validatedSyntheticVars
       cohortVariables = mkCohortVariables _qpiReferencedSessionVariables